requirements.txt
Django==1.11.2
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
PyJWT==1.5.0
pytz==2017.2
